Feature/consolidate (#3)
* feat: Add useDocumentMetadata hook for managing document metadata

The `useDocumentMetadata` hook is added to the `src/hooks` directory. This hook allows for dynamically updating the document title and adding meta tags to the document head. It takes a `title` parameter and an optional array of `metaTags` objects, each containing a `name` and `content` property. The hook sets the document title to the provided `title` and adds the specified meta tags to the document head. When the component using this hook is unmounted, the previous document title is restored and the added meta tags are removed from the document head.

This hook is used in the following files:
- `src/pages/ErrorPage.tsx`: Sets the document title to "Error - Trim Success".
- `src/pages/ForgotPasswordPage.tsx`: Sets the document title to "Contraseña Olvidada - Trim Success".
- `src/pages/LandingPage.tsx`: Sets the document title to "Trim Success".
- `src/pages/SignInPage.tsx`: Sets the document title to "Sign In - Trim Success".
- `src/pages/SignUpPage.tsx`: Sets the document title to "Sign Up - Trim Success".
- `src/pages/client/DashboardPage.tsx`: Sets the document title to "Panel - Trim Success".
- `src/pages/client/ImportPage.tsx`: Sets the document title to "Importar Datos - Trim Success".

This commit adds the `useDocumentMetadata` hook to the project, allowing for better management of document metadata.

* chore: Update FileUpload component to improve user experience

The FileUpload component in `src/components/pages/client/import/FileUpload.tsx` has been updated to improve the user experience. Changes include:
- Replacing the CircularProgress component with a loading button from the MUI lab.
- Adding icons for devices, storage, and forward to enhance visual appeal.
- Updating the text to inform users that the uploaded data will be used to create different reports.

This commit improves the FileUpload component to provide a better user experience.

* feat: Add react-imask dependency and PhoneField component

The `react-imask` dependency has been added to the project, with version `7.6.1`. This dependency allows for input masking in React applications. Additionally, a new `PhoneField` component has been added to the `src/components/common` directory. This component utilizes the `react-imask` library to provide a masked input field specifically for phone numbers.

This commit adds the `react-imask` dependency and the `PhoneField` component to the project.